(REF) APIv4 FieldSpec - Extract OptionsSpecTrait (options, optionsCallback)
